backend/requirements.txt
# FastAPI и ASGI server
fastapi==0.109.2
uvicorn[standard]==0.27.1
python-multipart==0.0.9

# Database
sqlalchemy==2.0.27
alembic==1.13.1
psycopg2-binary==2.9.9
asyncpg==0.29.0

# Redis
redis==5.0.1

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1

# Pydantic для settings и validation
pydantic==2.6.1
pydantic-settings==2.1.0

# Telegram
telethon==1.34.0
cryptography==42.0.2
python-telegram-bot==20.7

# LLM Integration
httpx==0.25.2
tenacity==8.2.3

# Scheduler
APScheduler==3.10.4

# Email
aiosmtplib==3.0.1
email-validator==2.3.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
httpx==0.25.2  # Already included above, but needed for TestClient
